requirements.txt
flask==3.0.0
openai==1.12.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
Pillow==10.2.0
xmltodict==0.13.0
flask-mail==0.9.1
werkzeug==3.0.1
gunicorn==21.2.0
